// drae-mobapp-app/src/types/profile.ts
export type PersonalInfo = {
  fullName: string;
  address: string;
  contactNumber: string;
  gender: string;
  age: string;
  email: string;
  contactPerson: string;
  contactPersonNumber: string;
  avatarUrl: string;
};

export const emptyPersonalInfo: PersonalInfo = {
  fullName: '',
  address: '',
  contactNumber: '',
  gender: '',
  age: '',
  email: '',
  contactPerson: '',
  contactPersonNumber: '',
  avatarUrl: '',
};